# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.FailureMsg:
    properties:
      error:
        type: string
    type: object
  model.MachineRole:
    properties:
      Id:
        type: integer
      creationDate:
        type: string
      creatorId:
        type: integer
      description:
        type: string
      machineRoleName:
        type: string
    type: object
...
  model.NetworkInterface:
    properties:
      Id:
        type: integer
      bitmask:
        type: integer
      creationDate:
        type: string
      creatorId:
        type: integer
      deviceId:
        type: string
      deviceModel:
        type: string
      gateway:
        type: string
      ipAddress:
        type: string
      macAddress:
        type: string
      systemId:
        type: integer
    type: object
  model.NetworkInterfaces:
    properties:
      interfaces:
        items:
          $ref: '#/definitions/model.NetworkInterface'
        type: array
    type: object
  model.OrgUnit:
    properties:
      Id:
        type: integer
      creationDate:
        type: string
      creatorId:
        type: integer
      description:
        type: string
      ouName:
        type: string
    type: object
  model.OrgUnitList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.OrgUnit'
        type: array
    type: object
...
  model.SuccessMsg:
    properties:
      message:
        type: string
    type: object
...
host: localhost:5000
info:
  contact:
    name: Gary Greene
    url: https://github.com/greeneg/allocatord
  description: An API for managing OS imaging
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Allocator Daemon
  version: 0.1.5
paths:
...
  /networkInterface:
    post:
      consumes:
      - application/json
      description: Add a new network interface
      parameters:
      - description: Network Interface data
        in: body
        name: networkInterface
        required: true
        schema:
          $ref: '#/definitions/model.NetworkInterface'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register network interface
      tags:
      - network-interfaces
  /networkInterface/{networkInterfaceId}:
    delete:
      consumes:
      - application/json
      description: Delete a network interface by Id
      parameters:
      - description: Network Interface Id
        in: path
        name: networkInterfaceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete network interface
      tags:
      - network-interfaces
    patch:
      description: Update a network interface by its Id
      parameters:
      - description: Network Interface ID
        in: path
        name: networkInterfaceId
        required: true
        type: integer
      - description: Network Interface data
        in: body
        name: networkInterfaceData
        required: true
        schema:
          $ref: '#/definitions/model.MachineRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Update a network interface by its Id
      tags:
      - network-interfaces
  /networkInterface/byId/{networkInterfaceId}:
    get:
      description: Retrieve a network interface by its Id
      parameters:
      - description: Network Interface ID
        in: path
        name: networkInterfaceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NetworkInterface'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a network interface by its Id
      tags:
      - network-interfaces
  /networkInterface/byIpAddress/{networkInterfaceIpAddress}:
    get:
      description: Retrieve a network interface by its IP Address
      parameters:
      - description: Network Interface IP Address
        in: path
        name: networkInterfaceIpAddress
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NetworkInterface'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a network interface by its IP Address
      tags:
      - network-interfaces
  /networkInterface/byMACAddress/{networkInterfaceMACAddress}:
    get:
      description: Retrieve a network interface by its MAC Address
      parameters:
      - description: Network Interface MAC Address
        in: path
        name: networkInterfaceMACAddress
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NetworkInterface'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve a network interface by its MAC Address
      tags:
      - network-interfaces
  /networkInterfaces:
    get:
      description: Retrieve list of all network interfaces
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NetworkInterfaces'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve list of all network interfaces
      tags:
      - network-interfaces
  /networkInterfaces/{systemId}:
    get:
      description: Retrieve the list of network interfaces for a system's Id
      parameters:
      - description: System ID
        in: path
        name: systemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NetworkInterfaces'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve the list of network interfaces for a system's Id
      tags:
      - network-interfaces
  /organizationalUnit:
    post:
      consumes:
      - application/json
      description: Add a new organizational unit
      parameters:
      - description: Organizational Unit data
        in: body
        name: orgUnit
        required: true
        schema:
          $ref: '#/definitions/model.OrgUnit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Register organizational unit
      tags:
      - orgs
  /organizationalUnit/{ouId}:
    delete:
      consumes:
      - application/json
      description: Delete an organizational unit
      parameters:
      - description: Organizational Unit Id
        in: path
        name: ouId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Delete OU
      tags:
      - orgs
  /organizationalUnit/byId/{ouId}:
    get:
      description: Retrieve an organizational unit by its Id
      parameters:
      - description: Organizationl Unit ID
        in: path
        name: ouId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrgUnit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve an organizational unit by its Id
      tags:
      - orgs
  /organizationalUnits:
    get:
      description: Retrieve list of all organizational units
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrgUnitList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.FailureMsg'
      security:
      - BasicAuth: []
      summary: Retrieve list of all organizational units
      tags:
      - orgs
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"